sqlcanon/stable-requirements.txt
#
# Project Requirements
#
MySQL-python==1.2.4
PIL==1.1.7
Django==1.4.3
pytz==2013b
South==0.7.6
sqlparse==0.1.6
mmh3==2.0
http://sourceforge.net/projects/pylibpcap/files/pylibpcap/0.6.4/pylibpcap-0.6.4.tar.gz/download
construct==2.5.1
django-crispy-forms==1.2.3
django-debug-toolbar==0.9.4
django-tastypie==0.9.15

#
# Other Dependencies
#
mimeparse==0.1.3
python-dateutil==1.5
six==1.3.0


